# backend/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: eCommerce
  description: >-
    Aplicación desacoplada de ventas online sobre Bases de Datos MongoDB
    - Rubro: vestimenta - 
    Provee las siguientes funcionalidades:
    1.- Organización de productos de acuerdo a categorias y subcategorías
    2.- Registro, logín y autenticación con credenciales locales y redes sociales 
    3.- Redes Sociales disponbles: Google, Facebook, Tweeter, GitHub, Instagram
    4.- Sistema de pagos automatizado
    5.- Galería de productos con carrusel de multiples imágenes
  contact:
    name: Camilo Morales Sanchez - Willy Escobar - Karem Urbano
    email: moorales.camilo@gmail.com 
servers:
  - url: http://localhost:8081/api/v1
...
paths:
  /cart:
    get:
      description: "Ruta para administración del carrito de compras"
      tags:
        - Cart
      security:
        - cookieAuth: []
      responses:
        200:
          description: "Los Items del carrito fueron obtenidos exitosamente"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listNoPagination"
        400:
          description: "Petición de solicitud de items de carrrito con errores"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: pedido de carrito rechazado por el servidor "
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Problemas con el servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /cart/clear:
    put:
      description: "Vacía el carrito"
      tags:
        - Cart
      security:
        - cookieAuth: []
      responses:
        200:
          description: "El carrito fue exitosamente vaciado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listNoPagination"
        400:
          description: "Error en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: Servodor rechaza el pedido"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Problemas con el servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /cart/add-item:
    post:
      description: "Agregar un item al carrito"
      tags:
        - Cart
      requestBody:
        description: "Datos del producto a agregar"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                  example: "62aa04ff6e95f9f27cdf6742"
                amount:
                  type: number
                  example: 1
      security:
        - cookieAuth: []
      responses:
        200:
          description: "Producto agregado al carrito satisfactoriamentee"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listNoPagination"
        400:
          description: "Producto no agregado por error en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: el servidor no puedo agregar el producto"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Problemas de servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /cart/remove-item:
    delete:
      description: "Eliminar un producto del carrito"
      tags:
        - Cart
      requestBody:
        description: "datos de producto a eliminar"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: string
                  example: "62aa04ff6e95f9f27cdf6742"
      security:
        - cookieAuth: []
      responses:
        200:
          description: "Producto satisfactoriamente eliminado del carrito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listNoPagination"
        400:
          description: "No se puede eliminar producto, petición mal realizada"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: inconvenientes del servidor para satisfacer lapetición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error de servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /cart/amount:
    put:
      description: "Cambia la cantidad de un producto del carrito"
      tags:
        - Cart
      requestBody:
        description: "Se le pasa el id del producto y la cantidad a modificar"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 1
                product:
                  type: string
                  example: "62aa04ff6e95f9f27cdf6742"
      security:
      security:
        - cookieAuth: []
      responses:
        200:
          description: "Cantidad de producto cambiada correctamente"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/listNoPagination"
        400:
          description: "No se puede moficar cantidad de producto del carrito por un error en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: el servidor no puede responder la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error de servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
...
  /signup/:
    post:
      description: "Ruta de Registro de Usuarios"
      tags:
        - Auth
      requestBody:
        description: "Alta de usuario en el sistema"
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        200:
          description: "Usuario registrado exitosamente"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseSuccessfulAuth"
        400:
          description: "Petición de registro erroneo"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /login/:
    post:
      tags:
        - Auth
      requestBody:
        description: "Inicio de sesión de Usuarios"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: correo@correo.com
                password:
                  type: string
                  example: Abc123**12
      security: []
      responses:
        200:
          description: "Usuario autorizado a ingresar al sistema"
          headers:
            Set-Cookie:
              schema:
                type: string
                example: cookie=abcde12345; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseSuccessfulAuth"

        400:
          description: "Pedido de autorización con errores"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        401:
          description: "Forbidden: Cuenta erronea"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
...
  /session/:
    get:
      description: "Sesión a través de Redes Sociales"
      tags:
        - Auth
      security:
        - cookieAuth: []
      responses:
        200:
          description: "Inicio de Sesión Con Redes Sociales Exitosa"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/responseSuccessfulAuth"
        400:
          description: "Petición de ingreso a traves de facebook rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        403:
          description: "Forbidden: Cuenta erronea"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /facebook/:
    get:
      description: "Obtención de credenciales en Facebook"
      tags:
        - SocialAuth
      responses:
        200:
          description: "Ingreso a través de credenciales de Facebook exitosa"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successful login"
        400:
          description: "Ingreso a traves de facebook rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /twitter/:
    get:
      description: "Obtención de credenciales en Twitter"
      tags:
        - SocialAuth
      responses:
        200:
          description: "Ingreso a través de credenciales de Twitter exitosa"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successful login"
        400:
          description: "Petición de ingreso a traves de Twitter rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /github/:
    get:
      description: "Obtención de credenciales en GitHub"
      tags:
        - SocialAuth
      responses:
        200:
          description: "Ingreso a través de credenciales de GitHub exitoso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successful login"
        400:
          description: "Petición de ingreso a traves de GitHub rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /instagram/:
    get:
      description: "Obtención de credenciales en Instagram"
      tags:
        - SocialAuth
      responses:
        200:
          description: "Ingreso a través de credenciales de Instagram exitosa"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successful login"
        400:
          description: "Petición de ingreso a través de Instagram rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /google/:
    get:
      description: "Obtención de credenciales en Google"
      tags:
        - SocialAuth
      responses:
        200:
          description: "Ingreso a través de credenciales de Google exitoso"
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Successful login"
        400:
          description: "Petición de ingreso a través de Google rechazada por errores en la petición"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        500:
          description: "Error interno de Servidor"
          content:
            aplication/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
...
components:
  schemas:
    user:
      type: object
      properties:
        name:
          type: string
          example: "name"
        email:
          type: string
          example: correo@correo.com
        role:
          type: integer
          example: 1
        password:
          type: string
          example: "Abc123**12"
        image:
          type: string
          format: binary
    responseSuccessfulAuth:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Lorem Ipsum is simply dummy text of the printing and typesetting
        name:
          type: string
          example: name
        email:
          type: string
          example: correo@correocom
        image:
          type: string
          example: http://localhost:8081/public/user/adba2f76-2a65-4566-bdae-9b8a0a7ac2c4.jpg
        role:
          type: integer
          example: 1
        id:
          type: string
          example: correo@correocom
    errorMessage:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              message:
                type: string
                example: "Lorem Ipsum is simply dummy text of the printing and typesetting"
...
    listNoPagination:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: Lorem Ipsum is simply dummy text of the printing and typesetting
        data:
          type: object
          properties:
            docs:
              type: array
              items:
                type: object
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token